src/Meridian.Scenarios/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
numpy>=1.26,<2
scipy>=1.12,<2
pydantic>=2.0,<3
pyodbc==5.*
python-dotenv==1.*
